# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateOut": {
        "description": "One row of the derivation trail (F4.AC11, DATA_MODEL section 5.4).",
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "admin1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin1"
          },
          "admin2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin2"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "effective_weight": {
            "title": "Effective Weight",
            "type": "number"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "raw_confidence": {
            "title": "Raw Confidence",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "suppressed_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suppressed Reason"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "source",
          "level",
          "country_code",
          "admin1",
          "admin2",
          "city",
          "lat",
          "lng",
          "raw_confidence",
          "weight",
          "effective_weight",
          "accepted",
          "suppressed_reason",
          "evidence",
          "latency_ms"
        ],
        "title": "CandidateOut",
        "type": "object"
      },
...
      "Classification": {
        "enum": [
          "human",
          "bot",
          "crawler",
          "datacenter",
          "spam",
          "spoofed",
          "unknown"
        ],
        "title": "Classification",
        "type": "string"
      },
...
      "DecryptedIp": {
        "properties": {
          "decrypted_at": {
            "format": "date-time",
            "title": "Decrypted At",
            "type": "string"
          },
          "ip": {
            "title": "Ip",
            "type": "string"
          }
        },
        "required": [
          "ip",
          "decrypted_at"
        ],
        "title": "DecryptedIp",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VisitDetail": {
        "properties": {
          "bot_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Score"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "classifier_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifier Version"
          },
          "client": {
            "additionalProperties": true,
            "title": "Client",
            "type": "object"
          },
          "consent_state": {
            "title": "Consent State",
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceBlock"
          },
          "finalized_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "geofence": {
            "$ref": "#/components/schemas/GeofenceBlock"
          },
          "honeypot_tripped": {
            "title": "Honeypot Tripped",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inference_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inference Version"
          },
          "inferred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inferred At"
          },
          "is_returning": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Returning"
          },
          "link": {
            "$ref": "#/components/schemas/LinkRef"
          },
          "location": {
            "$ref": "#/components/schemas/LocationBlock"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkBlock"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "referer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referer"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "signals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Signals",
            "type": "array"
          },
          "spoof_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spoof Score"
          },
          "stage": {
            "$ref": "#/components/schemas/VisitStage"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "utm": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm"
          },
          "visitor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visitor Id"
          }
        },
        "required": [
          "id",
          "occurred_at",
          "link",
          "stage",
          "classification",
          "bot_score",
          "spoof_score",
          "location",
          "network",
          "device",
          "geofence",
          "visitor_id",
          "is_returning",
          "finalized_at",
          "inferred_at",
          "trace_id",
          "classifier_version",
          "inference_version",
          "consent_state",
          "signals",
          "candidates",
          "client",
          "request",
          "referer",
          "utm",
          "honeypot_tripped"
        ],
        "title": "VisitDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Two surfaces sharing one deployment: a public capture surface (`/r/{slug}`) and an authenticated admin API (`/api/v1/*`). See docs/API.md.",
    "summary": "Visitor intelligence platform",
    "title": "Tracelet",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/visits/{visit_id}": {
      "get": {
        "operationId": "get_visit_api_v1_visits__visit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One visit, in full",
        "tags": [
          "visits"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/ip": {
      "get": {
        "description": "The only way a plaintext address leaves the system. Owner-only, limited to 10 an hour per admin, and it writes an audit row naming the actor and the visit before it answers (F12.AC4). 410 IP_PURGED past the TTL is expected, not a fault (ADR-0007).",
        "operationId": "decrypt_ip_api_v1_visits__visit_id__ip_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecryptedIp"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decrypt the address for one visit (owner only)",
        "tags": [
          "visits"
        ]
      }
    },
...
  }
}